Correct issues with switching from horiz to vert
The bug happened when the destination screen size did not allow the
specified configuration to fit. In particular, The example I ran into
is:
 * output 1 and 2 configured for 2560x1440
 * output 1 at 0x0
 * output 2 at 2560x0
 * screen size set to 5120x1440
When autorandr-rs atempted to configure the following setup:
 * output 1 at 0x1440
 * output 2 at 0x0
I ended up with both output 1 and 2 at 0x0 with panning across a virtual
desktop of size 5120x1440.

This happened because configuring output 1 failed, configuring
output 2 succeeded and the screen size did not change.

In this broken setup, I invoked xrandr(1) with a --dryrun command to see
what it would do in this situation.

I saw that it reconfigured the screen size. So I tried that too, and it
failed.

In debugging why it failed, I noticed that I was configuring a screen
size that would include allow the currently configured monitors. This
requires a more complicated sequence, a break-then-make sequence.

After this patch, autorandr-rs(1) now implements a break-then-make
sequence where it:
 1) disables all monitors that will be reconfigured, and any monitors
 that were enabled and are not part of the final setup
 2) sets the screen size
 3) re-enables and re-configures all monitors in the final setup

This seems to work reliably.

I also ran cargo(1) fmt and could not be bothered to separate that diff
from the aforementioned changes.

So