Skip screen size conf
Turns out autorandrd(1) was attempting to do a reconfigure that does
not make sense. Under some circumstances, for example when the current
screen resolution is a superset of the final resolution, autorandrd(1)
would reconfigure the screen an extra time, and sometimes this
reconfigure would fail. This failure would prevent autorandrd(1) from
finishing the reconfigure. This could happen several times in a row,
and exhaust all events from a hardware change.

This patch prevents all of that.